requirements.txt
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
tensorflow==2.1.0
tensorflow-hub==0.8.0
